stream/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-cors==0.7.0
aiokafka==0.11.0
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorlog==6.8.2
croniter==2.0.7
cryptography==43.0.3
faust-streaming==0.11.3
flaretool==0.2.3
frozenlist==1.4.1
future==1.0.0
idna==3.10
intervaltree==3.1.0
lark-parser==0.12.0
mode-streaming==0.4.1
multidict==6.1.0
mypy-extensions==1.0.0
ndg-httpsclient==0.5.1
opentracing==2.4.0
packaging==24.1
propcache==0.2.0
pusher==3.3.2
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyNaCl==1.5.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-whois==0.9.4
pytz==2024.2
requests==2.32.3
shortuuid==1.0.13
simplejson==3.19.3
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.6
terminaltables==3.1.10
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
venusian==3.1.0
yarl==1.15.5
z3-solver==4.13.4.0
